requirements.txt
pywebview>=5.0
bottle>=0.12.25
google-generativeai>=0.8.0
ezdxf>=1.1.0
python-dotenv>=1.0.0
